Move sign-up link to the right
Since sign-up and login are related, I've placed them next to
eachother: in the top right-hand side, where I think most sites
place them.

Since sign-up is also more important in the conversion path than
login, I've kept that to be the primary action. Log in is only
needed when the user logs out, and since it will be done more than
once, it can be learned.