Changing the user account deletion logic
Previously there was the possibility for user to delete their account
and to have their contributions linked to an anonymous account.

However, after discussion with our lawyer, we cannot sustain this
solution as we must be able to keep a certain amount of data for legal
reasons.

Currently this, commit gives the possibility for users to cancel their
subscribtion to the website and adds a "[USER LEFT}" to their username
so other people know this user isn't active anymore (a little bit like
Wikipedia).

I'm open to other solutions such as scrambling the username for example.
Mind that usernames must be unique so this could lead to certain issues.